+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:39807/16dfe5fc-2ab3-42ea-afa6-1d76c8c2b5d9?"}
+{"mongoUri":"mongodb://127.0.0.1:43977/fc94180c-86f8-46ad-b9e7-fd5803931b4d?"}
