+++ src/main/routes/signup-routes.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express'
 import { adaptRoute } from '../adapters/express-route-adapter'
-import { makeSignUpController } from '../factories/signup/signup'
+import { makeSignUpController } from '../factories/signup/signup-factory'
 
 export default (router: Router): void => {
   router.post('/signup', adaptRoute(makeSignUpController()))
