+++ src/presentation/controllers/signup/signup-controller.ts
@@ -4,7 +4,7 @@ import {
   Controller,
   AddAccount,
   Validation
-} from './signup-protocols'
+} from './signup-controller-protocols'
 import { badRequest, serverError, ok } from '../../helpers/http/http-helper'
 
 export class SignUpController implements Controller {
